python/determine_signal_data/test_determine_signal_data.py
import unittest
import numpy as np
import matplotlib.pyplot as plt
from determine_signal_data import plot_spectrum, identify_data_type

class TestFunctions(unittest.TestCase):
    def test_plot_spectrum(self):
        # Test with some sample data
        samples = np.random.rand(1000)
        sampling_rate = 1000
        plot_spectrum(samples, sampling_rate)
        # No assertion for plot, just check for exceptions

    def test_identify_data_type(self):
        # Test with audio sample data
        audio_samples = np.random.rand(1000)
        audio_sampling_rate = 44100
        identify_data_type(audio_samples, audio_sampling_rate)
        # No assertion for GUI updates, just check for exceptions

        # Test with image sample data
        image_samples = np.random.rand(1000)
        image_sampling_rate = 1000000
        identify_data_type(image_samples, image_sampling_rate)
        # No assertion for GUI updates, just check for exceptions

        # Test with unknown sample data
        unknown_samples = np.random.rand(1000)
        unknown_sampling_rate = 500000000
        identify_data_type(unknown_samples, unknown_sampling_rate)
        # No assertion for GUI updates, just check for exceptions

if __name__ == '__main__':
    unittest.main()
